update chamber behavioral sorting routine
instead of sorting pairs of partitions, we sort individual partitions

i.e. previously we would take partitions (1+2), (3+4), (5+6) and sort
for the best N from the "di-partition" collection of segments. Instead
the firmware now sorts each partition collection independently to choose
the best N.. this change is sensible with the introduction of
cross-partition ghost-cancellation (which is not yet implemented in the fw)