only recalculate the layer mask when required
previously the dynamic layer mask would override whatever is in the
firmware, negating the feature that the test bench calculates the layer
masks based on the firmware definition.

now if a global value is already defined it skips recalculation. May
also have some performance benefit depending on whether python
automatically memoizes this code